refactor(transforms): isolate compatibility normalization
